DPAD-1140 :: Add Ad integration
