+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@fandom/jwplayer-fandom",
-	"version": "2.0.0-canonical-ads-11",
+	"version": "2.0.0-canonical-ads-22",
 	"description": "JWPlayer for Fandom",
 	"exports": {
 		".": "./main.js",
