+++ src/jwplayer/players/CanonicalVideoPlayer/CanonicalVideoPlayer.tsx
@@ -1,10 +1,12 @@
-import React, { useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { CanonicalVideoPlayerProps } from 'jwplayer/types';
 import PlayerWrapper from 'jwplayer/players/shared/PlayerWrapper';
 import LoadableVideoPlayerWrapper from 'jwplayer/players/shared/LoadableVideoPlayerWrapper';
 import styled, { css } from 'styled-components';
 import useOnScreen from 'utils/useOnScreen';
 import useAdComplete from 'jwplayer/utils/useAdComplete';
+import { communicationService } from 'jwplayer/utils/communication';
+import { isLocalDevelopment, isOnBrowser } from 'jwplayer/utils/envs';
 
 const CanonicalVideoTopPlaceholder = styled.div`
 	width: 100%;
@@ -41,6 +43,16 @@ const CanonicalVideoPlayer: React.FC<CanonicalVideoPlayerProps> = ({ currentVide
 	const adComplete = useAdComplete();
 	const onScreen = useOnScreen(ref, '0px', 1);
 	const isScrollPlayer = !onScreen;
+
+	useEffect(() => {
+		const payload = {
+			siteType: 'web',
+			isProduction: isOnBrowser() && !isLocalDevelopment(),
+		};
+
+		// if the player's mounted communicate that the video platform is ready
+		communicationService.dispatch({ type: '[F2] Configured', payload });
+	}, []);
 
 	return (
 		<PlayerWrapper playerName="canonical-video-player">
